Remove hard-coded Class name:ID dicts
get_equipable_classes will also now fallback to the values
returned by get_equipped_classes since used_by_classes only
exists in TF2 and it stands to reason that an item is equipable
by a class if  that class already has it equipped.

The class ID map was a hack to workaround the fact that
used_by_classes uses string identifiers. Now there is little reason
to continue trying to use it.