Use the same HTTP user agent that the Steam client does
It's apparent at this point that some behavior is different for requests
made with the client (I'm also guessing JS behaves differently due
to a inclient var, but that is moot for this). Since we're
aiming to be as close to that as possible we might as well do it
this way.