Revert to previous is_name_prefixed generating behavior
It seems there are a few hardcoded cases in the game where items
are not prefixed that I cannot match, and the previous strict prefixing
was causing problems in TF2. As far as I can tell only a few unnoticable
cases like the haunted metal scrap will be affected by this behavior.