grequests/requests for HTTP, HTTP base class
This is a non-API breaking change that moves
steamodd from using pycurl to requests.

Not only does this make room to add fallback
code (which will make grequests and requests
optional dependencies) it is much easier to
write code for and does not have the security
issues that pycurl does, especially the issue
of it segfaulting randomly.

The HTTP base class is added for users who
want to use the same HTTP handler that steamodd
does and for implementing other format handlers
like XML and VDF.

Also added is the grequests compatibility patch.
This is a well tested fix to make grequests
compatible with the latest requests. The patch
is still waiting to be merged upstream. If you
wish to speak your mind about this go to
https://github.com/kennethreitz/grequests/pull/31