Handle pre-encoded strings in vdf. Fixes #3
Yes it's ugly and backwards, but this is the wonderful world of pre-3
string encoding. This is why the byte type was added, because people
think it's really weird that when they try to decode ascii shows up.
This is because it's trying to convert it into a byte sequence first.
But since there is no actual byte type in pre-3 python, thar you go.