Avoid (occasionally) harmful cast to int in attribute formatter
It seems that Valve has now began to, in addition to storing strings in
them, store float values in non-float schema fields. This messes things
up in cases where projects show attribute values without user overrides.
So we will only try to cast to int when the value isn't a float