Allow override of format and key params in steam.api
Now that Valve is introducing requests that return different results
depending on the key it is sometimes helpful to be able to override it
on a per-call basis. The usefulness of overriding the format is not yet
evident but who knows. Maybe valve will do things like mediawiki and
other places and have different kinds of json available