Add cells_used method
Why do we need this method?

I know that we can call it by len(inventory), but in some cases like using the Flask framework, we have to pass all arguments in the render_template(), so to pass the used slots we would have to create another argument, let's say 'cells_used=len(inventory)', something that can be avoided having a property in the items module to get the length. We have cells_total, why not have this one too?